[graph] ignore nodes whose outputs are not consumed
If the input graph contains a node (e.g. an initializer) whose output is
not consumed by any node, then the graph initialization fails.  More
specifically, in looking up the list of consumer nodes, the code indexes
into the `consumedby` dictionary, but the key (i.e. the output
name) does not exist in the dictionary since the output is unused.

This patch makes a quick fix by first checking whether the output name
is present in the dictionary before looking up the value in it.