Allow pathlib.Path values as input to Model construction
Prior to this patch, only string values could be used to specify the
path to the input ONNX file.  However, several Python programs use
values of type `pathlib.Path`, which causes the model creation to
silently fail in onnx-tool.  This patch fixes the problem by allowing
both string paths or `pathlib.Path` paths.

Credit to @xich for discovering the problem!